+++ packages/clerk-js/src/ui/customizables/makeCustomizable.tsx
@@ -6,7 +6,28 @@ import { appendEmojiSeparator, generateClassName } from './classGeneration';
 import type { ElementDescriptor, ElementId } from './elementDescriptors';
 
 type Customizable<T = Record<never, never>> = T & {
+  /**
+   * elementDescriptor - An optional property that can be an `ElementDescriptor` or an array of `ElementDescriptor` or `undefined`. This property is used to describe the elements for styling purposes.
+   *
+   * @example
+   * <Box
+   *   elementDescriptor={[descriptors.icon, descriptors.iconInitials]}
+   * />
+   *
+   * // Output: `cl-icon cl-iconInitials`
+   */
   elementDescriptor?: ElementDescriptor | Array<ElementDescriptor | undefined>;
+  /**
+   * elementId - An optional property that can be an `ElementId`. This property is used to set a unique identifier for the element.
+   *
+   * @example
+   * <Box
+   *   elementDescriptor={[descriptors.icon, descriptors.iconInitials]}
+   *   elementId={descriptors.icon.setId(id)}
+   * />
+   *
+   * // Output: `cl-icon cl-iconInitials cl-icon__google cl-iconInitials__google`
+   */
   elementId?: ElementId;
   css?: never;
   sx?: ThemableCssProp;